Moved localization files to suitable dirs
