+++ src/localization/localizations/de/components/navbar.ts
