+++ src/localization/localizations/de/de.ts
@@ -1,9 +1,9 @@
-import { homepage } from '@/localization/localizations/de/homepage'
-import { register } from '@/localization/localizations/de/register'
-import { login } from '@/localization/localizations/de/login'
-import { profile } from '@/localization/localizations/de/profile'
-import { navbar } from '@/localization/localizations/de/navbar'
-import { errors } from '@/localization/localizations/de/errors'
+import { homepage } from '@/localization/localizations/de/views/homepage'
+import { register } from '@/localization/localizations/de/views/register'
+import { login } from '@/localization/localizations/de/views/login'
+import { profile } from '@/localization/localizations/de/views/profile'
+import { navbar } from '@/localization/localizations/de/components/navbar'
+import { errors } from '@/localization/localizations/de/views/errors'
 
 export const deutsch = {
   homepage,
