+++ src/localization/localizations/de/views/errors.ts
