+++ src/localization/localizations/de/views/homepage.ts
