+++ src/localization/localizations/de/views/register.ts
