+++ src/localization/localizations/en/components/navbar.ts
