+++ src/localization/localizations/en/en.ts
@@ -1,9 +1,9 @@
-import { homepage } from '@/localization/localizations/en/homepage'
-import { register } from '@/localization/localizations/en/register'
-import { login } from '@/localization/localizations/en/login'
-import { profile } from '@/localization/localizations/en/profile'
-import { navbar } from '@/localization/localizations/en/navbar'
-import { errors } from '@/localization/localizations/en/errors'
+import { homepage } from '@/localization/localizations/en/views/homepage'
+import { register } from '@/localization/localizations/en/views/register'
+import { login } from '@/localization/localizations/en/views/login'
+import { profile } from '@/localization/localizations/en/views/profile'
+import { navbar } from '@/localization/localizations/en/components/navbar'
+import { errors } from '@/localization/localizations/en/views/errors'
 
 export const english = {
   homepage,
