+++ src/localization/localizations/en/views/errors.ts
