+++ src/localization/localizations/en/views/homepage.ts
