+++ src/localization/localizations/en/views/register.ts
